feat: Remove controller area, refactor index
