+++ src/Dfe.PlanTech.Domain/Core/Enums/Pages.cs
@@ -1,0 +1,6 @@
+namespace Dfe.PlanTech.Domain.Core.Enums;
+
+public enum Pages
+{
+    Landing
+}
